Aplicacion del Single Responsability Principle
Se separo las responsabilidades de la clase SongDescriptionHelperImpl
Se añadio la clase DateUtils para el manejo de funciones relacionadas a fechas.
Se añadio la clase ReleaseDateBuilder para separar la responsabilidad de generar la release date.
Ademas se dividio la funcion getSongDescriptionText para separar la obtencion de la descripcion del armado de la misma.
